fix Not Found error failures from GraphQL client
